Refactor System list table to use Inventory toolbar
