--- src/Components/SmartComponents/SystemsPage/SystemTableToolbar.test.js
@@ -1,182 +1,0 @@
-import SystemsTableToolbar from './SystemsTableToolbar';
-import toJson from 'enzyme-to-json';
-import { mountWithIntl } from '../../../Helpers/MiscHelper';
-import { BrowserRouter as Router } from 'react-router-dom';
-import configureStore from 'redux-mock-store';
-import { Provider } from 'react-redux';
-import { intl } from '../../../Utilities/IntlProvider';
-import useOsVersionFilter from '../../PresentationalComponents/Filters/PrimaryToolbarFilters/OsVersionFilter';
-import { systems } from '../../../Helpers/fixtures';
-
-const mockStore = configureStore([store => next => action => { }]);
-let store = mockStore({});
-
-const applyMock = jest.fn();
-const handleSelectMock = jest.fn();
-const doOptOutMock = jest.fn();
-
-jest.mock('../../PresentationalComponents/Filters/PrimaryToolbarFilters/OsVersionFilter')
-
-let wrapper;
-
-beforeEach(() => {
-    wrapper = mountWithIntl(
-        <Provider store={store}>
-            <Router>
-                <SystemsTableToolbar
-                    intl={intl}
-                    methods={{
-                        apply: applyMock,
-                        handleSelect: handleSelectMock,
-                        doOptOut: doOptOutMock
-                    }}
-                    selectedRows={[]}
-                    canExport
-                    canSetExcludedIncluded
-                    canReadExcluded
-                    parameters={{
-                        excluded: "false",
-                        page: 1,
-                        page_size: 20,
-                        sort: "-updated"
-                    }}
-                    rawData={{ data: systems.data, meta: { totalItems: 10 } }}
-                />
-            </Router>
-        </Provider>
-    );
-});
-
-describe('SystemsTableToolbar component', () => {
-    it("Renders toolbar", () => {
-        expect(
-          wrapper.find({
-            ouiaId: "PrimaryToolbar",
-          })
-        ).toHaveLength(1);
-      });
-
-    it('Should match snapshot', () => {
-        expect(toJson(wrapper.find('Toolbar'))).toMatchSnapshot();
-    });
-
-    it('Select all systems', () => {
-        const selectAllToggle = wrapper.find('DropdownToggle button').first();
-        selectAllToggle.simulate('click');
-
-        wrapper.find('Dropdown').first().find("button").at(2).simulate('click');
-        const bulkSelectedSystems = systems.data.map((system) => {
-            return {
-                display_name: system.display_name,
-                id: system.id,
-                opt_out: system.attributes.opt_out,
-                culled_timestamp: system.culled_timestamp,
-                cve_count: system.cve_count,
-                insights_id: system.insights_id,
-                last_evaluation: system.last_evaluation,
-                last_upload: system.last_upload,
-                os: system.os,
-                rules_evaluation: system.rules_evaluation,
-                stale_timestamp: system.stale_timestamp,
-                stale_warning_timestamp: system.stale_warning_timestamp,
-                tags: system.tags,
-                type: system.type,
-                updated: system.updated,
-                remediation: system.remediation,
-                selected: true
-            };
-        });
-
-        expect(handleSelectMock).toHaveBeenCalledWith(bulkSelectedSystems, true);
-
-    });
-
-    it('Should exclude the selected system', () => {
-        wrapper = mountWithIntl(
-            <Provider store={store}>
-                <Router>
-                    <SystemsTableToolbar
-                        intl={intl}
-                        methods={{
-                            apply: applyMock,
-                            handleSelect: handleSelectMock,
-                            doOptOut: doOptOutMock
-                        }}
-                        selectedRows={[{ display_name: "hypervisor.beav", id: "ece17ad9-97db-4480-a486-3eb4edd6c330", opt_out: false, remediation: undefined, selected: true }]}
-                        selectedRowsCount={1}
-                        totalSelectedRows={1}
-                        canExport
-                        canSetExcludedIncluded
-                        canReadExcluded
-                        parameters={{
-                            excluded: "false",
-                            page: 1,
-                            page_size: 20,
-                            sort: "-updated"
-                        }}
-                        rawData={{ data: systems.data, meta: { totalItems: 10 } }}
-                    />
-                </Router>
-            </Provider>
-        );
-
-        const actionsToggle = wrapper.find('KebabToggle button');
-        actionsToggle.simulate('click');
-
-        const exclude = wrapper.find('Dropdown').at(3).find("DropdownItem button").at(0)
-        const include = wrapper.find('Dropdown').at(3).find("DropdownItem button").at(1);
-
-        exclude.simulate('click');
-
-        expect(include.props('disabled')).toBeTruthy();
-        expect(exclude.text()).toBe('Disable Vulnerability analysis on system');
-        expect(doOptOutMock).toHaveBeenCalled();
-    })
-
-    it('Should include the excluded selected system', () => {
-        wrapper = mountWithIntl(
-            <Provider store={store}>
-                <Router>
-                    <SystemsTableToolbar
-                        intl={intl}
-                        methods={{
-                            apply: applyMock,
-                            handleSelect: handleSelectMock,
-                            doOptOut: doOptOutMock
-                        }}
-                        selectedRows={[{ display_name: "client01.apex.example.com", id: "84b763c8-9611-4bf0-9fb8-4f737253b994", opt_out: true, remediation: undefined, selected: true }]}
-                        selectedRowsCount={1}
-                        canExport
-                        canSetExcludedIncluded
-                        canReadExcluded
-                        parameters={{
-                            excluded: "false",
-                            page: 1,
-                            page_size: 20,
-                            sort: "-updated"
-                        }}
-                        rawData={{ data: systems.data, meta: { totalItems: 10 } }}
-                    />
-                </Router>
-            </Provider>
-        );
-
-        const actionsToggle = wrapper.find('KebabToggle button');
-        actionsToggle.simulate('click');
-
-        const exclude = wrapper.find('Dropdown').at(3).find("DropdownItem button").at(0)
-        const include = wrapper.find('Dropdown').at(3).find("DropdownItem button").at(1);
-
-        include.simulate('click');
-
-        expect(exclude.props('disabled')).toBeTruthy();
-        expect(include.text()).toBe('Enable Vulnerability analysis on system');
-        expect(doOptOutMock).toHaveBeenCalled();
-    })
-
-    it("The default filter is set by name", () => {
-        const container = wrapper.find('button[data-ouia-component-id="ConditionalFilter"]');
-        expect(container).toHaveLength(1);
-        expect(container.text()).toBe('Name');
-      })
-});
