+++ src/Components/SmartComponents/SystemsPage/SystemsPage.js
@@ -1,9 +1,8 @@
-import React, { useEffect, useState, Fragment } from 'react';
+import React, { useEffect, useState, Fragment, useMemo } from 'react';
 import { useIntl } from 'react-intl';
 import messages from '../../../Messages';
 import { useDispatch, useSelector, shallowEqual } from 'react-redux';
-import SystemsTableToolbar from './SystemsTableToolbar';
-import { PERMISSIONS, SYSTEMS_ALLOWED_PARAMS } from '../../../Helpers/constants';
+import { PERMISSIONS, SYSTEMS_ALLOWED_PARAMS, SYSTEMS_DEFAULT_FILTERS, SYSTEMS_FILTER_PARAMS } from '../../../Helpers/constants';
 import ReducerRegistry from '../../../Utilities/ReducerRegistry';
 import { Main } from '@redhat-cloud-services/frontend-components/Main';
 import { systemTableRowActions } from '../../../Helpers/CVEHelper';
@@ -14,36 +13,39 @@ import {
     clearSystemStore,
     clearInventoryStore,
     selectRows,
-    changeColumnsSystemList
+    changeColumnsSystemList,
+    fetchSystems,
+    fetchSystemsIds
 } from '../../../Store/Actions/Actions';
 import { useUrlParams } from '../../../Helpers/MiscHelper';
 import { InventoryTable } from '@redhat-cloud-services/frontend-components/Inventory';
 import ErrorHandler from '../../PresentationalComponents/ErrorHandler/ErrorHandler';
 import { TableVariant } from '@patternfly/react-table';
-import { useColumnManagement, useGetEntities, useOptOutSystems, useRbac } from '../../../Helpers/Hooks';
+import { useBulkSelect, useColumnManagement, useGetEntities, useOptOutSystems, useRbac } from '../../../Helpers/Hooks';
 import * as APIHelper from '../../../Helpers/APIHelper';
 import { EmptyStateNoSystems } from '../../PresentationalComponents/EmptyStates/EmptyStates';
-import { clearNotifications } from '@redhat-cloud-services/frontend-components-notifications/redux';
+import { addNotification, clearNotifications } from '@redhat-cloud-services/frontend-components-notifications/redux';
 import NoAccessPage from '../../PresentationalComponents/StaticPages/NoAccessPage';
 import Spinner from '@redhat-cloud-services/frontend-components/Spinner';
+import { buildActiveFilters, isFilterInDefaultState, removeFilters } from '../../../Helpers/TableToolbarHelper';
+import DownloadReport from '../../../Helpers/DownloadReport';
+import useOsVersionFilter from '../../PresentationalComponents/Filters/PrimaryToolbarFilters/OsVersionFilter';
+import useSearchFilter from '../../PresentationalComponents/Filters/PrimaryToolbarFilters/SearchFilter';
+import excludedFilter from '../../PresentationalComponents/Filters/PrimaryToolbarFilters/ExcludedFilter';
 
 const SystemsPage = () => {
-    const [[canReadVulnerabilityResults,
+    const [[
+        canReadVulnerabilityResults,
         canSetExcludedIncluded,
-        canExport,
-        canReadExcluded],
-    isLoading] = useRbac([
+        canReadExcluded
+    ], isLoading
+    ] = useRbac([
         PERMISSIONS.readVulnerabilityResults,
         PERMISSIONS.setExcludedIncluded,
-        PERMISSIONS.basicReporting,
         PERMISSIONS.readExcluded
     ]);
 
-    const [[
-        canReadHostsInventory],
-    isLoadingInventory] = useRbac([
-        PERMISSIONS.readHosts
-    ], 'inventory');
+    const [[canReadHostsInventory], isLoadingInventory] = useRbac([PERMISSIONS.readHosts], 'inventory');
 
     const inventoryRef = React.createRef();
     const dispatch = useDispatch();
@@ -90,12 +92,10 @@ const SystemsPage = () => {
 
     const handleSelect = (payload, selecting) => dispatch(selectRows(payload, selecting));
 
-    const onRefreshInventory = () => {
-        dispatch(clearInventoryStore());
-        // timestamp is used to force inventory to refresh
-        // if it wasn't there inventory might ignore request to refresh because parameters are the same
-        inventoryRef.current.onRefreshData(({ timestamp: Date.now(), page: 1 }));
-    };
+    const onRefreshInventory = () => (
+        dispatch(clearInventoryStore()),
+        inventoryRef.current.onRefreshData(({ page: 1 }))
+    );
 
     const doOptOut = useOptOutSystems(onRefreshInventory);
     const getEntities = useGetEntities(APIHelper.getSystems, {});
@@ -103,17 +103,82 @@ const SystemsPage = () => {
     const [columnCounter, setColumnCount] = useState(0);
     useEffect(() => setColumnCount(columnCounter + 1), [columns]);
 
-    // TODO: let InventoryTable render its own toolbar instead of using custom one
-    return (
-        isLoading ? <Spinner centered/> :
-            canReadVulnerabilityResults ? <Fragment>
-                { ColumnManagementModal }
+    const downloadReport = format => {
+        let params = { ...parameters };
+        DownloadReport.exec(
+            fetchSystems,
+            params,
+            format,
+            'system-list',
+            notification => dispatch(
+                addNotification(notification)
+            ),
+            () => dispatch(clearNotifications())
+        );
+    };
 
-                <Header title={intl.formatMessage(messages.vulnerabilitySystemsHeader)} showBreadcrumb={false}/>
+    const kebabProps = useMemo(() => {
+        return {
+            selectedExcluded: selectedRows.some(({ opt_out: optOut }) => optOut === true),
+            selectedIncluded: selectedRows.some(({ opt_out: optOut }) => optOut === false)
+        };
+    }, [selectedRows]);
+
+    const kebabOptions = [
+        '',
+        ...canSetExcludedIncluded ? [{
+            label: intl.formatMessage(messages.systemKebabDisableAnalysis, { count: selectedRowsCount }),
+            onClick: () => doOptOut(selectedRows, selectedRows?.[0].display_name, true),
+            props: { isDisabled: !selectedRowsCount || !kebabProps.selectedIncluded }
+        },
+        {
+            label: intl.formatMessage(messages.systemKebabEnableAnalysis, { count: selectedRowsCount }),
+            onClick: () => doOptOut(selectedRows, selectedRows?.[0].display_name, false, selectedRows),
+            props: { isDisabled: !selectedRowsCount || !kebabProps.selectedExcluded }
+        }] : [],
+        {
+            label: intl.formatMessage(messages.columnManagementModalTitle),
+            onClick: () => setColumnManagementModalOpen(true)
+        }
+    ];
+
+    const rawData = { data: items, meta: { totalItems }, isLoaded };
+
+    const bulkSelectProps = useBulkSelect({
+        rawData,
+        selectedRows,
+        selectedRowsCount,
+        handleSelect,
+        fetchResource: ops => fetchSystemsIds({ ...parameters, ...ops })
+    });
+
+    const osVersionFilter = useOsVersionFilter(
+        parameters.rhel_version,
+        apply
+    );
+
+    const filterConfigItems = [
+        useSearchFilter(
+            'filter',
+            messages.systemsSearchName,
+            messages.searchFilterByName,
+            parameters.filter,
+            apply
+        ),
+        ...(canReadExcluded ? [excludedFilter(apply, parameters)] : []),
+        ...osVersionFilter
+    ];
+
+    return (
+        isLoading ? <Spinner centered /> :
+            canReadVulnerabilityResults ? <Fragment>
+                {ColumnManagementModal}
+
+                <Header title={intl.formatMessage(messages.vulnerabilitySystemsHeader)} showBreadcrumb={false} />
                 <Main>
-                    {isLoadingInventory ? <Spinner centered/> :
+                    {isLoadingInventory ? <Spinner centered /> :
                         hasError && !canReadHostsInventory
-                            ? <ErrorHandler code={errorCode}/>
+                            ? <ErrorHandler code={errorCode} />
                             : (
                                 <InventoryTable
                                     disableDefaultColumns
@@ -157,28 +222,30 @@ const SystemsPage = () => {
                                     columns={defaultColumns => mergeColumns(defaultColumns)}
                                     getEntities={getEntities}
                                     hideFilters={{ all: true }}
-                                    noSystemsTable={<EmptyStateNoSystems/>}
-                                >
-                                    <SystemsTableToolbar
-                                        canExport={canExport}
-                                        canSetExcludedIncluded={canSetExcludedIncluded}
-                                        canReadExcluded={canReadExcluded}
-                                        parameters={parameters}
-                                        rawData={{ data: items, meta: { totalItems }, isLoaded }}
-                                        selectedRows={selectedRows}
-                                        selectedRowsCount={selectedRowsCount}
-                                        methods={{
-                                            doOptOut,
-                                            apply,
-                                            handleSelect,
-                                            setColumnManagementModalOpen
-                                        }}
-                                        actions
-                                    />
-                                </InventoryTable>
+                                    showTags
+                                    noSystemsTable={<EmptyStateNoSystems />}
+                                    bulkSelect={bulkSelectProps}
+                                    actionsConfig={{
+                                        actions: kebabOptions,
+                                        dropdownProps: { ouiaId: 'toolbar-actions' }
+                                    }}
+                                    filterConfig={{
+                                        items: filterConfigItems
+                                    }}
+                                    activeFiltersConfig={{
+                                        filters: buildActiveFilters(parameters),
+                                        onDelete: (_, chips, reset) =>
+                                            removeFilters(chips, apply, reset, SYSTEMS_DEFAULT_FILTERS),
+                                        deleteTitle: intl.formatMessage(messages.resetFilters),
+                                        showDeleteButton: !isFilterInDefaultState(
+                                            parameters,
+                                            canReadExcluded ? SYSTEMS_DEFAULT_FILTERS : {},
+                                            SYSTEMS_FILTER_PARAMS)
+                                    }}
+                                />
                             )}
                 </Main>
-            </Fragment> : <NoAccessPage/>
+            </Fragment> : <NoAccessPage />
     );
 };
